BF config.combat.heroic_strike_damage is now 176 (rank 10), down from 208 (rank 11, which isn't available) BF base_special_stats now properly reflect level 70 values (190 instead of 160 AP, and 1.141% instead of 0% crit); added a test for that ENH/BF draenei will now have buffs.heroic_presence automatically enabled ENH added or improved output operators for Special_stats, Attributes, and Race ENH added proper crit detection for rampage ("rampage aura"); both overpower and rampage aura are now only used where necessary, and don't use the buff manager anymore ENH scheduling via time (e.g. time + cooldown) instead of duration (e.g. cooldown, repeatedly "increment()ed" until it's <= 0) ENH removed some "constexpr" from functions which are, in fact, no const expressions at all ENH/BF over time buffs no longer need to occur on a fixed timestamp;  handling is vastly simplified, and allows refreshes for any kind of over time effect; as a result, Deep Wounds damage is calculated more accurately ENH improved dps output for tests, "optimized" Damage_sources::add_damage() (buff-manager "re-write" work-in-progress)

+++ wow_library/include/Item.hpp
@@ -148,7 +148,7 @@ public:
     Hit_effect() = default;
 
     Hit_effect(std::string name, Type type, Attributes attribute_boost, Special_stats special_stats_boost,
-               double damage, int duration, double cooldown, double probability, double attack_power_boost = 0, int n_targets = 1,
+               double damage, double duration, double cooldown, double probability, double attack_power_boost = 0, int n_targets = 1,
                int armor_reduction = 0, int max_stacks = 0, double ppm = 0.0, bool affects_both_weapons = false)
         : name(std::move(name))
         , type(type)
@@ -175,7 +175,7 @@ public:
     Attributes attribute_boost;
     Special_stats special_stats_boost;
     double damage;
-    int duration;
+    double duration;
     double cooldown;
     double probability;
     double attack_power_boost;
@@ -216,6 +216,7 @@ public:
         , hit_effects(std::move(hit_effects))
         , over_time_effects(std::move(over_time_effects)){};
 
+    // FIXME ap_multiplier is used for attributes, but not for special_stats...
     inline Special_stats get_special_stat_equivalent(const Special_stats& special_stats, double ap_multiplier = 0) const
     {
         return attribute_boost.convert_to_special_stats(special_stats, ap_multiplier) + special_stats_boost;
